Playing a game completed
I've added the API routes for:
- Getting a player's hand
- Getting the tricks that have been played so far (including the current
  trick)
- Playing a card

Some error validation is present. There is still a few things that need
doing such as scoring and starting the whole game again.

Also haven't tested all of the logic yet, just a few combinations on the
first trick. I'd like to add some unit tests to work through the different
behaviors but this can come after I've got scoring going (enough for the
clients to start to be devloped).