Converted Game to Pure Functions, added Mocha tests
I had to refactor the game engine so that bits were reusable in clients,
and so had to break up a lot of the validation functions into pure
functions that can be called from the client, not relying on an object
that is in the correct state. This also made my logic easier to test.

I've added Mocha tests, one of them is failing at the moment but I'm tired
and so have stopped. I'll get it working as the next thing I do.

The HAPI api now provides a reason why a play has been rejected. It isn't
specific enough yet to tell the user what was wrong, but it will tell you
the difference between not being your turn, not having the card you tried
to play, and the game deciding the card is illegal.