Add SemanticUI and some simple AI
SemanticUI added as a CSS library to make the client slightly more
appealing.

A simple AI has been added for all players except player 0, the AI chooses
the first valid card it has in its hand. Note this is quite dumb and
should be improved at some point :)

Also added in a .gitattributes file to deal with the line endings conflict
between windows and unix.